+++ src/main/javascript/CustomFields/index.js
@@ -1,0 +1,6 @@
+/**
+ * CustomFields module.
+ * @module CustomFields
+ */
+
+export { CustomFieldsClient } from './CustomFieldsClient';